新增接口，删除银行卡信息

+++ app/Http/Controllers/Api/Module/Organization/BankController.php
@@ -159,6 +159,12 @@ class BankController extends BaseController
       $where = ['id' => $request->id];
 
       $model = $this->_model::getRow($condition);
+
+      if(empty($model->id))
+      {
+        return self::error(Code::ERROR);
+      }
+
       $model->delete();
 
       return self::success(Code::message(Code::HANDLE_SUCCESS));
